ai_service/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
opencv-python-headless
numpy
boto3
python-dotenv
uuid
